Refactor some patches out of redis.spec
For cases where we have two ways of doing one thing (using
a patch or sed-based modification of files), rework these
such that we're only using one method.  Allows several of
the patches to be dropped.

Described in https://github.com/antirez/redis/issues/1417
is a method for allowing integration/replication-psync to
be reliably run - start using it, and drop the patch that
was disabling that test.

Other small side-effects of this commit:
- the build warning about "jemalloc not found" is fixed;
- CFLAGS additions are now consistently handled (for deps
  as well) - using the upstream-provided mechanism for
  compiler flag propogation.