Incorporate recent feedback from Remi, build once
Drop modification of the configuration file to add
systemd supervision - Remi correctly points out this
is redundant (due to unit files already overriding)
and will just cause issues.

Noticed the build runs twice also, once during the
%build phase and once again during %install (full
rebuild) - this was because of differing options on
the make lines (CFLAGS= etc vs PREFIX= etc) and the
Redis src/Makefile causes full rebuild as a result
(including of deps/* which it "make distclean"s).

Also fixed a rpmbuild "warning: Macro expanded in
comment" message in a couple of places, and changed
the way we squash hidden build messages to use the
"V" macro as the Makefiles intended.