Redis v4 modules packaging preparation
Add spec file macros to create a default location for
Redis modules, and to allow module packages to depend
on a specific redis module ABI.  A build time test is
added to ensure the upstream sources API version does
not differ to one the redis package Provides:.

The devel package now provides an rpm macros file for
the %{redis_modules_abi} and %{redis_modules_dir}.

Much of this is ideas from Haikel and Remi - thanks!

Small cleanup - the _hardened_build macro is dropped
as this has been the default since f22.