Addition of a redis-doc subpackage for Redis v4 onward
This adds a redis-doc subpackage, containing the upstream
docs (https://github.com/antirez/redis-doc/), as well as
the existing man page content we've added.  Care is taken
to handle the movement of man page files between packages
and the naming convention of the redis-doc source is also
careful to follow the prefered Fedora conventions.

Optional conversion of markdown to html is done by pandoc
if available.

In the future, it would be to modify the redis build to
generate the redis src/help.h file using the content from
redis-doc now that we have it.  It's often out of date, I
am noticing - seems to be updated in an ad-hoc fashion in
upstream releases.